nix: switch to meson build system + override stdenv to remove -no_uuid ldflag.
On macOS, nix by default uses an ldflag named "-no_uuid" via bintools ->
clang-wrapper -> stdenv.cc -> stdenv to improve build reproducibility but
unfortunately, lldb on darwin also relies on build UUIDs for symbolication and
source mapping.

Therefore, as we are debugging sigtool, we override bintools, clang-wrapper,
and stdenv to remove this flag [and while we're at it, we idiosyncratically
port the sigtool build system from make/cmake to meson.