codesign: use mkstemp() filename output to drive codesign_allocate
1. Rather than synthesizing a path to the tempfile created by `mkstemp(char*
template)` via `/dev/fd/...`, we instead use the fact that `mkstemp()` mutates
the buffer pointed to by `template` to directly obtain the name of the
generated tempfile.

2. Ignore EINTR while waitpid()'ing for codesign_allocate to finish. (This, +
changes to the buildsystem in another commit enable us to use lldb to debug
failing executions.)

3. Finally, while we're at it, also stop `Commands::codesign()` from leaking
`tempfileName` via a suitable std::unique_ptr.